backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
pyyaml==6.0.1
python-dotenv==1.0.0

# 现有依赖（复用 spreadsheet-normalizer）
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
numpy>=1.24.0
openai>=1.0.0
chardet>=5.0.0
